// superset-frontend/plugins/superset-plugin-chart-custom-district-map/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { QueryFormData } from '@superset-ui/core';

export interface SupersetPluginChartCustomDistrictMapStylesProps {
  height: number;
  width: number;
}

interface SupersetPluginChartCustomDistrictMapCustomizeProps {
  selectState: string;
  entity: string;
  numberFormat: string;
  linearColorScheme: string;
  showDistrictLabels: boolean;
}

export type SupersetPluginChartCustomDistrictMapQueryFormData = QueryFormData &
  SupersetPluginChartCustomDistrictMapStylesProps &
  SupersetPluginChartCustomDistrictMapCustomizeProps;

export interface DistrictMapDataItem {
  district_id: string;
  metric: number;
}

export type SupersetPluginChartCustomDistrictMapProps =
  SupersetPluginChartCustomDistrictMapStylesProps &
    SupersetPluginChartCustomDistrictMapCustomizeProps & {
      data: DistrictMapDataItem[];
      sliceId: number;
      setDataMask: (dataMask: Record<string, unknown>) => void;
      filterState: { selectedValues?: string[] | null };
      emitCrossFilters: boolean;
    };